Flesh out HTTP/File unit tests
